refactor: 거리,시간 계산 유틸 common->hub 이동
