+++ hub/src/main/java/com/klp/hub/hub/util/DistanceTimeUtil.java
@@ -1,4 +1,4 @@
-package com.klp.hub.common.util;
+package com.klp.hub.hub.util;
 
 public final class DistanceTimeUtil {
